backend/requirements.txt
# API Framework
fastapi==0.95.0
uvicorn==0.21.1

# Database
sqlalchemy==2.0.7
psycopg2-binary==2.9.6
databases==0.7.0
alembic==1.10.2

# Authentication & Security
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
cryptography==41.0.1
python-multipart==0.0.6

# Data Validation
pydantic==1.10.7
email-validator==2.0.0.post2

# HTTP and Services
requests==2.28.2
httpx==0.24.1

# Environment and Configuration
python-dotenv==1.0.0

# Logging and Monitoring
loguru==0.7.0

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0

# Documentation
markupsafe==2.1.2
